+++ OneBlink.SDK/models/FormElementAttributeMapping.cs
@@ -1,0 +1,10 @@
+using System;
+
+namespace OneBlink.SDK.Model
+{
+    public class FormElementAttributeMapping
+    {
+        public Guid elementId { get; set; }
+        public string @attribute { get; set; }
+    }
+}